chore: relocate `isRenderableElement` to `services/render`
