+++ src/services/index.ts
@@ -1,5 +1,3 @@
-import * as Logging from 'hyperview/src/services/logging';
-import * as Namespaces from 'hyperview/src/services/namespaces';
 import * as Xml from 'hyperview/src/services/xml';
 import { DEFAULT_PRESS_OPACITY, HV_TIMEOUT_ID_ATTR } from './types';
 import type {
@@ -8,7 +6,7 @@ import type {
   StyleSheet,
   StyleSheets,
 } from 'hyperview/src/types';
-import { LOCAL_NAME, NODE_TYPE } from 'hyperview/src/types';
+import { NODE_TYPE } from 'hyperview/src/types';
 import { Platform } from 'react-native';
 
 /**
@@ -287,101 +285,6 @@ export const uuid = (): string => {
  */
 export const uuidNumber = (): number => {
   return parseInt(uuid().replace(/-/g, ''), 16);
-};
-
-/**
- * Checks if an element should be rendered based on the element type and options
- * Logging warnings for unexpected conditions
- * @param {Element} element
- * @param {HvComponentOptions} options
- * @param {InlineContext} inlineFormattingContext
- * @returns {boolean}
- */
-export const isRenderableElement = (
-  element: Element,
-  options: HvComponentOptions,
-  inlineFormattingContext: [Node[], string[]] | null | undefined,
-): boolean => {
-  if (!element) {
-    return false;
-  }
-  if (
-    element.nodeType === NODE_TYPE.ELEMENT_NODE &&
-    element.getAttribute('hide') === 'true'
-  ) {
-    // Hidden elements don't get rendered
-    return false;
-  }
-  if (element.nodeType === NODE_TYPE.COMMENT_NODE) {
-    // XML comments don't get rendered.
-    return false;
-  }
-  if (
-    element.nodeType === NODE_TYPE.ELEMENT_NODE &&
-    element.namespaceURI === Namespaces.HYPERVIEW
-  ) {
-    switch (element.localName) {
-      case LOCAL_NAME.BEHAVIOR:
-      case LOCAL_NAME.MODIFIER:
-      case LOCAL_NAME.STYLES:
-      case LOCAL_NAME.STYLE:
-        // Non-UI elements don't get rendered
-        return false;
-      default:
-        break;
-    }
-  }
-
-  if (element.nodeType === NODE_TYPE.ELEMENT_NODE) {
-    if (!element.namespaceURI) {
-      Logging.warn('`namespaceURI` missing for node:', element.toString());
-      return false;
-    }
-    if (!element.localName) {
-      Logging.warn('`localName` missing for node:', element.toString());
-      return false;
-    }
-
-    if (
-      options.componentRegistry?.getComponent(
-        element.namespaceURI,
-        element.localName,
-      )
-    ) {
-      // Has a component registered for the namespace/local name.
-      return true;
-    }
-    // No component registered for the namespace/local name.
-    // Warn in case this was an unintended mistake.
-    Logging.warn(
-      `No component registered for tag <${element.localName}> (namespace: ${element.namespaceURI})`,
-    );
-  }
-
-  if (element.nodeType === NODE_TYPE.TEXT_NODE) {
-    // Render non-empty text nodes, when wrapped inside a <text> element
-    if (element.nodeValue) {
-      if (
-        ((element.parentNode as Element)?.namespaceURI ===
-          Namespaces.HYPERVIEW &&
-          (element.parentNode as Element)?.localName === LOCAL_NAME.TEXT) ||
-        (element.parentNode as Element)?.namespaceURI !== Namespaces.HYPERVIEW
-      ) {
-        if (options.preformatted) {
-          return true;
-        }
-        // When inline formatting context exists, lookup formatted value using node's index.
-        if (inlineFormattingContext) {
-          return true;
-        }
-      }
-    }
-  }
-
-  if (element.nodeType === NODE_TYPE.CDATA_SECTION_NODE) {
-    return true;
-  }
-  return false;
 };
 
 /**
