+++ src/services/render/index.tsx
@@ -1,4 +1,6 @@
 import * as InlineContext from 'hyperview/src/services/inline-context';
+import * as Logging from 'hyperview/src/services/logging';
+import * as Namespaces from 'hyperview/src/services/namespaces';
 import type {
   HvComponentOnUpdate,
   HvComponentOptions,
@@ -8,7 +10,101 @@ import type {
 import { LOCAL_NAME, NODE_TYPE } from 'hyperview/src/types';
 import HvElement from 'hyperview/src/components/hv-element';
 import React from 'react';
-import { isRenderableElement } from 'hyperview/src/services';
+
+/**
+ * Checks if an element should be rendered based on the element type and options
+ * Logging warnings for unexpected conditions
+ * @param {Element} element
+ * @param {HvComponentOptions} options
+ * @param {InlineContext} inlineFormattingContext
+ * @returns {boolean}
+ */
+export const isRenderableElement = (
+  element: Element,
+  options: HvComponentOptions,
+  inlineFormattingContext: [Node[], string[]] | null | undefined,
+): boolean => {
+  if (!element) {
+    return false;
+  }
+  if (
+    element.nodeType === NODE_TYPE.ELEMENT_NODE &&
+    element.getAttribute('hide') === 'true'
+  ) {
+    // Hidden elements don't get rendered
+    return false;
+  }
+  if (element.nodeType === NODE_TYPE.COMMENT_NODE) {
+    // XML comments don't get rendered.
+    return false;
+  }
+  if (
+    element.nodeType === NODE_TYPE.ELEMENT_NODE &&
+    element.namespaceURI === Namespaces.HYPERVIEW
+  ) {
+    switch (element.localName) {
+      case LOCAL_NAME.BEHAVIOR:
+      case LOCAL_NAME.MODIFIER:
+      case LOCAL_NAME.STYLES:
+      case LOCAL_NAME.STYLE:
+        // Non-UI elements don't get rendered
+        return false;
+      default:
+        break;
+    }
+  }
+
+  if (element.nodeType === NODE_TYPE.ELEMENT_NODE) {
+    if (!element.namespaceURI) {
+      Logging.warn('`namespaceURI` missing for node:', element.toString());
+      return false;
+    }
+    if (!element.localName) {
+      Logging.warn('`localName` missing for node:', element.toString());
+      return false;
+    }
+
+    if (
+      options.componentRegistry?.getComponent(
+        element.namespaceURI,
+        element.localName,
+      )
+    ) {
+      // Has a component registered for the namespace/local name.
+      return true;
+    }
+    // No component registered for the namespace/local name.
+    // Warn in case this was an unintended mistake.
+    Logging.warn(
+      `No component registered for tag <${element.localName}> (namespace: ${element.namespaceURI})`,
+    );
+  }
+
+  if (element.nodeType === NODE_TYPE.TEXT_NODE) {
+    // Render non-empty text nodes, when wrapped inside a <text> element
+    if (element.nodeValue) {
+      if (
+        ((element.parentNode as Element)?.namespaceURI ===
+          Namespaces.HYPERVIEW &&
+          (element.parentNode as Element)?.localName === LOCAL_NAME.TEXT) ||
+        (element.parentNode as Element)?.namespaceURI !== Namespaces.HYPERVIEW
+      ) {
+        if (options.preformatted) {
+          return true;
+        }
+        // When inline formatting context exists, lookup formatted value using node's index.
+        if (inlineFormattingContext) {
+          return true;
+        }
+      }
+    }
+  }
+
+  if (element.nodeType === NODE_TYPE.CDATA_SECTION_NODE) {
+    return true;
+  }
+  return false;
+};
 
 export const renderElement = (
   element: Element | null | undefined,
